requirements.txt
absl-py==2.0.0
altair==5.2.0
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
distro==1.8.0
flatbuffers==23.5.26
fonttools==4.46.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
mediapipe==0.10.8
numpy==1.24.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==14.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pypdfium2==4.25.0
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
streamlit==1.29.0
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
